SLING-8627 - Update sling-jcr-repoinit to Oak 1.16.0 and Jackrabbit 2.18.2
Update the compile and IT dependencies to the latest Oak and Jackrabbit versions.

To make sure we get the right dependencies, we now include all the oak artefacts
in a custom oak.txt provisioning file. We also changed the name of the main
feature to Sling, to make sure overrides are working.